+++ memote/support/helpers.py
@@ -158,6 +158,9 @@ def find_transport_reactions(model):
     transport_reactions = []
     transport_rxn_candidates = set(model.reactions) - set(model.boundary) \
         - set(find_biomass_reaction(model))
+    transport_rxn_candidates = set(
+        [rxn for rxn in transport_rxn_candidates if len(rxn.compartments) >= 2]
+    )
     # Add all labeled transport reactions
     sbo_matches = set([rxn for rxn in transport_rxn_candidates if
                        rxn.annotation is not None and